\docType{package}
\name{rk.MPT-package}
\alias{rk.MPT}
\alias{rk.MPT-package}
\title{RKWard GUI for multinomial processing tree (MPT) models.}
\description{
  RKWard GUI for multinomial processing tree (MPT) models.
}
\details{
  \tabular{ll}{ Package: \tab rk.MPT\cr Type: \tab
  Package\cr Version: \tab 0.01-2\cr Date: \tab
  2012-04-07\cr Depends: \tab rkward (>= 0.5.6)\cr
  Enhances: \tab rkward\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab http://rkward.sf.net\cr }

  RKWard GUI for multinomial processing tree (MPT) models
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

